Extract diagnostics/E2E commands (oh-tab-o9sx)
